Bei dieser Version muss mit sh compile.sh kompiliert werden. Die compile.sh Datei befinted sich im Root-Verzeichnis des Projektordners.
