+++ compile.sh
@@ -1,0 +1,8 @@
+#g++ -g -o server Server.cpp MasterCube.cpp
+#g++ -g -o client Client.cpp MasterCube.cpp
+
+g++ -g -o main src/main.cpp src/masterCube.cpp src/client.cpp src/server.cpp
+
+# Makefile muss noch geschrieben werden
+# ./server
+# ./client
