+++ src/client.h
@@ -1,0 +1,1 @@
+int SetupClient();
