+++ src/makefile
@@ -171,12 +171,12 @@ ETAGS = etags
 CTAGS = ctags
 am__DIST_COMMON = $(srcdir)/makefile.in $(top_srcdir)/depcomp
 DISTFILES = $(DIST_COMMON) $(DIST_SOURCES) $(TEXINFOS) $(EXTRA_DIST)
-ACLOCAL = ${SHELL} /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/missing aclocal-1.15
+ACLOCAL = ${SHELL} ../mpk_AIproject_group3/missing aclocal-1.15
 AMTAR = $${TAR-tar}
 AM_DEFAULT_VERBOSITY = 1
-AUTOCONF = ${SHELL} /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/missing autoconf
-AUTOHEADER = ${SHELL} /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/missing autoheader
-AUTOMAKE = ${SHELL} /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/missing automake-1.15
+AUTOCONF = ${SHELL} ../mpk_AIproject_group3/missing autoconf
+AUTOHEADER = ${SHELL} ../mpk_AIproject_group3/missing autoheader
+AUTOMAKE = ${SHELL} ../mpk_AIproject_group3/missing automake-1.15
 AWK = mawk
 CC = gcc
 CCDEPMODE = depmode=gcc3
@@ -201,7 +201,7 @@ LDFLAGS =
 LIBOBJS = 
 LIBS = 
 LTLIBOBJS = 
-MAKEINFO = ${SHELL} /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/missing makeinfo
+MAKEINFO = ${SHELL} ../mpk_AIproject_group3/missing makeinfo
 MKDIR_P = /bin/mkdir -p
 OBJEXT = o
 PACKAGE = mastercube
@@ -216,10 +216,10 @@ SET_MAKE =
 SHELL = /bin/bash
 STRIP = 
 VERSION = 1.0
-abs_builddir = /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/src
-abs_srcdir = /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/src
-abs_top_builddir = /home/grooty/Documents/mpk_ai/mpk_AIproject_group3
-abs_top_srcdir = /home/grooty/Documents/mpk_ai/mpk_AIproject_group3
+abs_builddir = ../mpk_AIproject_group3/src
+abs_srcdir = ../mpk_AIproject_group3/src
+abs_top_builddir = ../mpk_AIproject_group3
+abs_top_srcdir = ../mpk_AIproject_group3
 ac_ct_CC = gcc
 ac_ct_CXX = g++
 am__include = include
@@ -239,7 +239,7 @@ host_alias =
 htmldir = ${docdir}
 includedir = ${prefix}/include
 infodir = ${datarootdir}/info
-install_sh = ${SHELL} /home/grooty/Documents/mpk_ai/mpk_AIproject_group3/install-sh
+install_sh = ${SHELL} ../mpk_AIproject_group3/install-sh
 libdir = ${exec_prefix}/lib
 libexecdir = ${exec_prefix}/libexec
 localedir = ${datarootdir}/locale
