+++ src/server.h
@@ -1,0 +1,1 @@
+void SetupServer();
